requirements.txt
python3-discogs-client
pyyaml
requests
beautifulsoup4
lxml
schedule
pyTelegramBotAPI
